Looping with for Go back to your controller. Eventually, users will need to be able to comment on the articles, so they can respectfully debate the article's conclusions based on objective analysis and research. Ya know... no different than any other news commenting section. Ahem.
I'll paste in 3 fake comments. Add a second variable called comments to pass these into the template. This time, we can't just print that array: we need to loop over it. At the bottom, and an h2 that says 'Comments' and then add a ul. To loop, we need our first do something tag! Woo! Use {% for comment in comments %}. Most 'do' something tags also have a closing tag: {% endfor %}.Inside the loop, comment represents the individual comment. So, just print it: {{ comment }}.Try it! Brilliant! I mean, it's really ugly... oof. But we'll fix that later.