Template Inheritance Twig has one last killer feature: it's template inheritance system. Because remember! We don't yet have a real HTML page: just the content from the template.
To fix this, at the top of the template, add {% extends 'base.html.twig' %}
This refers to the base.html.twig file that was added by the recipe.

It's simple now, but this is our layout file and we'll customize it over time. By extending it, we should at least get this basic HTML structure.

But when we refresh... surprise! An error! And probably one that you'll see at some point!

A template that extends another one cannot include content outside Twig blocks