Unpacking the Pack! Go back to your Twig template and remove that dump. Then, open composer.json. We just installed two packs: the debug-pack and the profiler-pack:
 67 lines  composer.json
{
... lines 2 - 15
    require-dev: {
... line 17
        'symfony/debug-pack': '^1.0',
... line 19
        'symfony/profiler-pack': '^1.0'
    },
... lines 22 - 65
}
And we now know that the debug-pack is actually a collection of about 6 libraries.

But, packs have a disadvantage... a 'dark side'. What if you wanted to control the version of just one of these libraries? Or what if you wanted most of these libraries, but you didn't want, for example, the phpunit-bridge. Well... right now, there's no way to do that: all we have is this one debug-pack line.

Don't worry brave space traveler! Just... unpack the pack! Yep, at your terminal, run:

composer unpack debug

The unpack command comes from Symfony flex. And... interesting! All it says is 'removing symfony/debug-pack'.
But look at your composer.json. Ah! It did remove symfony/debug-pack, but it replaced it with the 6 libraries from that pack! We can now control the versions or even remove individual libraries if we don't want them.

That is the power of packs!