The Not-So-Mystical asset Function And now I'm going to slightly complicate things. Go back into PhpStorm's Preferences, search for 'Symfony' and find the 'Symfony' plugin. Change the 'web' directory to public - it was called web in Symfony 3.
This is not required, but it will give us more auto-completion when working with assets. Delete the 'font-awesome' path, re-type it, and hit tab to auto-complete.
Woh! It wrapped the path in a Twig asset() function! Do the same thing below for styles.css

Here's the deal: whenever you link to a static asset - CSS, JS or images - you should wrap the path in this asset() function. But... it's not really that important. In fact, right now, it doesn't do anything: it will print the same path as before. But! In the future, the asset() function will give us more flexibility to version our assets or store them on a CDN.

In other words: don't worry about it too much, but do remember to use it!