Let's take the dynamic code that we have at the bottom and work it into the new HTML. For the title, use {{ title }}
Below, it prints the number of comments. Replace that with {{ comments|length }}

Oh, and at the bottom, there is a comment box and one actual comment. Let's find this and... add a loop! For comment in comments on top, and endfor at the bottom. For the actual comment, use {{ comment }}

Delete the old code from the bottom... oh, but don't delete the endblock.

Let's try it - refresh! It looks awesome! A bunch of things are still hardcoded, but this is much better.

It's time to make our homepage less ugly and learn about the second job of routing: route generation for linking.