Most of these links don't got anywhere yet. Whatever! No problem! We're going to fill them in as we continue. Besides, most of our users will be in hypersleep for at least a few more decades.
But we can hook up some of these - like the 'Space Bar' logo text - that should go to the homepage.

Open templates/base.html.twig and search for 'The Space Bar'

Ok - let's point this link to the homepage. And yep, we could just say href='/'.

But... there's a better way. Instead, we're going to generate a URL to the route. Yep, we're going to ask Symfony to give us the URL to the route that's above our homepage action.

Why? Because if we ever decided to change this route's URL - like to /news - if we generate the URL instead of hardcoding it, all the links will automatically update. Magic!

The Famous debug:router
So how can we do this? First, find your terminal and run:

./bin/console debug:router
This is an awesome little tool that shows you a list of all of the routes in your app. You can see our two routes and a bunch of routes that help the profiler and web debug toolbar.

There's one thing about routes that we haven't really talked about yet: each route has an internal name. This is never shown to the user, it only exists so that we can refer to that route in our code. For annotation routes, by default, that name is created for us.

Generating URLs with path()
This means, to generate a URL to the homepage, copy the route name, go back to base.html.twig, add {{ path() }} and paste the route name.
That's it!

Refresh! Click it! Yes! We're back on the homepage.