This route has a wildcard... so we can't just generate a URL to it. Nope, we need to also tell Symfony what value it should use for the {slug} part.
How? Add a second argument to path(): {}. That's the syntax for an associative array when you're inside Twig - it's similar to JavaScript. Give this a slug key set to why-asteroids-taste-like-bacon.

Try it - refresh! Error gone! And check this out: the link goes to our show page.

Next, let's add some JavaScript and an API endpoint to bring this little heart icon to life!