Making the Route POST-Only Eventually, this endpoint will modify something on the server - it will like the article. So as a best-practice, we should not be able to make a GET request to it. Let's make this route only match when a POST request is made. How? Add another option to the route: methods={'POST'}
As soon as we do that, we can no longer make a GET request in the browser: it does not match the route anymore! Run:

./bin/console debug:router
And you'll see that the new route only responds to POST requests. Pretty cool. By the way, Symfony has a lot more tools for creating API endpoints - this is just the beginning. In future tutorials, we'll go further!