Hooking up the JavaScript & API Our API endpoint is ready! Copy the route name and go back to article_show.js. But wait... if we want to make an AJAX request to the new route... how can we generate the URL? This is a pure JS file... so we can't use the Twig path() function!
Actually, there is a really cool bundle called FOSJsRoutingBundle(https://github.com/FriendsOfSymfony/FOSJsRoutingBundle) that does allow you to generate routes in JavaScript. But, I'm going to show you another, simple way.

Back in the template, find the heart section. Let's just... fill in the href on the link! Add path(), paste the route name, and pass the slug wildcard set to a slug variable.

Actually... there is not a slug variable in this template yet. If you look at ArticleController, we're only passing two variables. Add a third: slug set to {dollar_sign}slug.

That should at least set the URL on the link. Go back to the show page in your browser and refresh. Yep! The heart link is hooked up.