Why did we do this? Because now we can get that URL really easily in JavaScript. Add {dollar_sign}.ajax({}) and pass method: 'POST' and url set to {dollar_sign}link.attr('href'). That's it! At the end, add .done() with a callback that has a data argument. The data will be whatever our API endpoint sends back. That means that we can move the article count HTML line into this, and set it to data.heart.
Go Deeper!

Oh, and if you're not familiar with the .done() function or Promises, I'd highly recommend checking out our JavaScript Track. It's not beginner stuff: it's meant to take your JS up to the next level.

Anyways... let's try it already! Refresh! And... click! It works!

And... I have a surprise! See this little arrow icon in the web debug toolbar? This showed up as soon as we made the first AJAX request. Actually, every time we make an AJAX request, it's added to the top of this list! That's awesome because - remember the profiler? - you can click to view the profiler for any AJAX request. Yep, you now have all the performance and debugging tools at your fingertips... even for AJAX calls.

Oh, and if there were an error, you would see it in all its beautiful, styled glory on the Exception tab. Being able to load the profiler for an AJAX call is kind of an easter egg: not everyone knows about it. But you should.

I think it's time to talk about the most important part of Symfony: Fabien. I mean, services.