Using Twig Directly And check this out! If you want to get the Twig service, you can use either of these two type-hints.
And remember how I said that everything in Symfony is done by a service? Well, when we call {dollar_sign}this->render() in a controller, that's just a shortcut to fetch the Twig service and call a method on it.

In fact, let's pretend that the ->render() shortcut does not exist. How could we render a template? No problem: we just need the Twig service. Add a second argument with an Environment type-hint, because that's the class name we saw in debug:autowiring. Call the arg twigEnvironment. Next, change the return statement to be {dollar_sign}html = {dollar_sign}twigEnvironment->render().

The method we want to call on the Twig object is coincidentally the same as the controller shortcut.

Then at the bottom, return new Response() and pass html.

Ok, this is way more work than before... and I would not do this in a real project. But, I wanted to prove a point: when you use the {dollar_sign}this->render() shortcut method on the controller, all it really does is call render() on the Twig service and then wrap it inside a Response object for you.

Try it! Go back and refresh the page. It works exactly like before! Of course we will use shortcut methods, because they make our life way more awesome.